miAPIJWT/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]   # Para JWT
passlib[bcrypt]             # Para hashear contraseñas